+++ src/main/java/org/goafabric/calleeservice/logic/CalleeLogic.java
@@ -1,32 +1,36 @@
 package org.goafabric.calleeservice.logic;
 
+import lombok.NonNull;
 import lombok.SneakyThrows;
+import org.goafabric.calleeservice.service.Callee;
 import org.springframework.stereotype.Component;
+
+import java.util.concurrent.atomic.AtomicLong;
 
 @Component
 public class CalleeLogic {
-    private Long sleepTime = 0l;
+    private AtomicLong sleepTime = new AtomicLong(0L);
 
-    public Callee sayMyName(String name) {
+    public Callee sayMyName(@NonNull String name) {
         sleep();
         return Callee.builder()
                 .message("Your name is: " + name).build();
     }
 
-    public Callee sayMyOtherName(String name) {
+    public Callee sayMyOtherName(@NonNull String name) {
         sleep();
         return Callee.builder()
                 .message("Your name is: " + name).build();
     }
 
-    public Callee setSleepTime(Long sleepTime) {
-        this.sleepTime = sleepTime; //you should never change instance variables for a productive app
+    public Callee setSleepTime(@NonNull Long sleepTime) {
+        this.sleepTime.set(sleepTime); //you should never change instance variables for a productive app
         return Callee.builder()
                 .message("set sleepTime to: " + sleepTime).build();
     }
 
     @SneakyThrows
     private void sleep() {
-        Thread.sleep(sleepTime);
+        Thread.sleep(sleepTime.get());
     }
 }